+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/RatifySpec.hs
@@ -626,7 +626,7 @@ votingSpec =
           getLastEnactedCommittee `shouldReturn` SNothing
           -- Bump up the UTxO delegated
           -- to barely make the threshold (65 %! 100)
-          stakingKP1 <- lookupKeyPair stakingKH1
+          stakingKP1 <- getKeyPair stakingKH1
           sendCoinTo_ (mkAddr paymentKP1 stakingKP1) (inject $ Coin 858_000_000)
           passNEpochs 2
           -- The same vote should now successfully ratify the proposal
@@ -673,13 +673,13 @@ votingSpec =
           (drepKH1, stakingKH1) <- setupDRepWithoutStake
           -- Add rewards to delegation #1
           submitAndExpireProposalToMakeReward $ KeyHashObj stakingKH1
-          lookupReward (KeyHashObj stakingKH1) `shouldReturn` govActionDeposit
+          getReward (KeyHashObj stakingKH1) `shouldReturn` govActionDeposit
           -- Setup DRep delegation #2
           (_drepKH2, stakingKH2) <- setupDRepWithoutStake
           (spoC, _, _) <- setupPoolWithStake $ Coin 42_000_000
           -- Add rewards to delegation #2
           submitAndExpireProposalToMakeReward $ KeyHashObj stakingKH2
-          lookupReward (KeyHashObj stakingKH2) `shouldReturn` govActionDeposit
+          getReward (KeyHashObj stakingKH2) `shouldReturn` govActionDeposit
           -- Submit a committee proposal
           cc <- KeyHashObj <$> freshKeyHash
           Positive extra <- arbitrary
@@ -695,7 +695,7 @@ votingSpec =
           -- Increase the rewards of the delegator to this DRep
           -- to barely make the threshold (65 %! 100)
           registerAndRetirePoolToMakeReward $ KeyHashObj stakingKH1
-          lookupReward (KeyHashObj stakingKH1) `shouldReturn` poolDeposit <> govActionDeposit
+          getReward (KeyHashObj stakingKH1) `shouldReturn` poolDeposit <> govActionDeposit
           isDRepAccepted addCCGaid `shouldReturn` True
           -- The same vote should now successfully ratify the proposal
           passEpoch
@@ -991,12 +991,12 @@ votingSpec =
             (poolKH1, delegatorCStaking1) <- setupPoolWithoutStake
             -- Add rewards to delegation #1
             submitAndExpireProposalToMakeReward delegatorCStaking1
-            lookupReward delegatorCStaking1 `shouldReturn` govActionDeposit
+            getReward delegatorCStaking1 `shouldReturn` govActionDeposit
             -- Setup Pool delegation #2
             (poolKH2, delegatorCStaking2) <- setupPoolWithoutStake
             -- Add rewards to delegation #2
             submitAndExpireProposalToMakeReward delegatorCStaking2
-            lookupReward delegatorCStaking2 `shouldReturn` govActionDeposit
+            getReward delegatorCStaking2 `shouldReturn` govActionDeposit
             -- Submit a committee proposal
             Positive extra <- arbitrary
             cc <- KeyHashObj <$> freshKeyHash
@@ -1017,7 +1017,7 @@ votingSpec =
             -- Add to the rewards of the delegator to this SPO
             -- to barely make the threshold (51 %! 100)
             registerAndRetirePoolToMakeReward delegatorCStaking1
-            lookupReward delegatorCStaking1 `shouldReturn` poolDeposit <> govActionDeposit
+            getReward delegatorCStaking1 `shouldReturn` poolDeposit <> govActionDeposit
             -- The same vote should now successfully ratify the proposal
             -- NOTE: It takes 2 epochs for SPO votes as opposed to 1 epoch
             -- for DRep votes to ratify a proposal.